mise à jour

+++ lib/controller/financePartener.dart
@@ -38,18 +38,18 @@ class financepartenaireState extends State<financePartenaire> {
   ];
 
   bool showAvg = false;
-  int ancienjan = 0;
-  int ancienfev = 0;
-  int ancienmar = 0;
-  int ancienavril = 0;
-  int ancienmai = 0;
-  int ancienjuin = 0;
-  int ancienjuil = 0;
-  int ancienaout = 0;
-  int anciensep = 0;
-  int ancienoct = 0;
-  int anciennov = 0;
-  int anciendec = 0;
+  double ancienjan = 0;
+  double ancienfev = 0;
+  double ancienmar = 0;
+  double ancienavril = 0;
+  double ancienmai = 0;
+  double ancienjuin = 0;
+  double ancienjuil = 0;
+  double ancienaout = 0;
+  double anciensep = 0;
+  double ancienoct = 0;
+  double anciennov = 0;
+  double anciendec = 0;
 
 
   //
@@ -164,7 +164,7 @@ class financepartenaireState extends State<financePartenaire> {
   }
 
   String ValeurTotal(chiffres business){
-    int total=business.janvier+business.fevrier+business.mars+business.avril+business.mai+business.juin+business.juillet+business.aout+business.septembre+business.octobre+business.novembre+business.decembre;
+    double total=business.janvier+business.fevrier+business.mars+business.avril+business.mai+business.juin+business.juillet+business.aout+business.septembre+business.octobre+business.novembre+business.decembre;
     return formatchiffre.format(total).toString();
 
   }
@@ -173,7 +173,7 @@ class financepartenaireState extends State<financePartenaire> {
   LineChartData mainData(chiffres business) {
     int decimals=0;
     int fac=pow(10,decimals);
-    int total=business.janvier+business.fevrier+business.mars+business.avril+business.mai+business.juin+business.juillet+business.aout+business.septembre+business.octobre+business.novembre+business.decembre;
+    double total=business.janvier+business.fevrier+business.mars+business.avril+business.mai+business.juin+business.juillet+business.aout+business.septembre+business.octobre+business.novembre+business.decembre;
     double moyenne=total/12;
     moyenne=(moyenne*fac).round()/fac;
     return LineChartData(
@@ -497,7 +497,7 @@ class financepartenaireState extends State<financePartenaire> {
 
 
   Widget graphfromage(chiffres business){
-    int total=business.janvier+business.fevrier+business.mars+business.avril+business.mai+business.juin+business.juillet+business.aout+business.septembre+business.octobre+business.novembre+business.decembre;
+    double total=business.janvier+business.fevrier+business.mars+business.avril+business.mai+business.juin+business.juillet+business.aout+business.septembre+business.octobre+business.novembre+business.decembre;
     return AspectRatio(
       aspectRatio: 1.6,
       child: Card(
@@ -637,7 +637,7 @@ class financepartenaireState extends State<financePartenaire> {
   }
 
 
-  List<PieChartSectionData> showingSections(chiffres business,int total) {
+  List<PieChartSectionData> showingSections(chiffres business,double total) {
     int decimals=1;
     int fac=pow(10,decimals);
     double jan=(business.janvier)/total*100;
